read_xpt {haven}	R Documentation

Read and write SAS transport files

Description

The SAS transport format is a open format, as is required for submission of the data to the FDA.

Usage

read_xpt(file)

write_xpt(data, path, version = 8, name = NULL)

Arguments

file
Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. It must contain at least one new line to be recognised as data (instead of a path).

data
Data frame to write.

path
Path to a file where the data will be written.

version
Version of transport file specification to use: either 5 or 8.

name
Member name to record in file. Defaults to file name sans extension. Must be <= 8 characters for version 5, and <= 32 characters for version 8.

Value

A tibble, data frame variant with nice defaults.

Variable labels are stored in the "label" attribute of each variable. It is not printed on the console, but the RStudio viewer will show it.

write_xpt() returns the input data invisibly.

Examples

tmp <- tempfile(fileext = ".xpt")
write_xpt(mtcars, tmp)
read_xpt(tmp)

[Package haven version 1.1.2 Index]