zap_formats {haven}	R Documentation

Remove format attributes

Description

To provide some mild support for round-tripping variables between Stata/SPSS and R, haven stores variable formats in an attribute: format.stata, format.spss, or format.sas. If this causes problems for your code, you can get rid of them with zap_formats.

Usage

zap_formats(x)

Arguments

x
A vector or data frame.

See Also

Other zappers: zap_empty(), zap_labels(), zap_label(), zap_widths()

[Package haven version 2.4.3 Index]