zap_widths {haven}	R Documentation

Remove display width attributes

Description

To provide some mild support for round-tripping variables between SPSS and R, haven stores display widths in an attribute: display_width. If this causes problems for your code, you can get rid of them with zap_widths.

Usage

zap_widths(x)

Arguments

x
A vector or data frame.

See Also

Other zappers: zap_empty(), zap_formats(), zap_labels(), zap_label()

[Package haven version 2.4.3 Index]